requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alabaster==0.7.12         # via sphinx
appdirs==1.4.3            # via black, virtualenv
asgiref==3.2.7            # via django
assertpy==1.0             # via -r requirements.in
astunparse==1.6.3         # via sphinxcontrib-httpexample
attrs==19.3.0             # via black, pytest
autoflake==1.3.1          # via -r requirements.in
babel==2.8.0              # via sphinx
black==19.10b0            # via -r requirements.in
boto3==1.12.26            # via -r requirements.in
botocore==1.15.31         # via boto3, s3transfer
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.1.1              # via black
distlib==0.3.0            # via virtualenv
django-anymail==7.0.0     # via -r requirements.in
django-cors-headers==3.2.1  # via -r requirements.in
django-filter==2.2.0      # via -r requirements.in
django==3.0.4             # via -r requirements.in, django-anymail, django-cors-headers, django-filter, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
docutils==0.15.2          # via botocore, sphinx, sphinxcontrib-httpexample
envparse==0.2.0           # via -r requirements.in
factory-boy==2.12.0       # via -r requirements.in
faker==4.0.2              # via factory-boy
filelock==3.0.12          # via tox, virtualenv
freezegun==0.3.15         # via -r requirements.in
idna==2.9                 # via requests
imagesize==1.2.0          # via sphinx
invoke==1.4.1             # via -r requirements.in
isort==4.3.21             # via -r requirements.in
jinja2==2.11.2            # via sphinx
jmespath==0.9.5           # via boto3, botocore
markupsafe==1.1.1         # via jinja2
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest, sphinx, tox
pathspec==0.7.0           # via black
pillow==7.1.1             # via -r requirements.in
pluggy==0.13.1            # via pytest, tox
psycopg2-binary==2.8.4    # via -r requirements.in
py==1.8.1                 # via pytest, tox
pyflakes==2.2.0           # via autoflake
pygments==2.6.1           # via sphinx
pyparsing==2.4.6          # via packaging
pytest-django==3.8.0      # via -r requirements.in
pytest==5.4.1             # via -r requirements.in, pytest-django
python-dateutil==2.8.1    # via botocore, faker, freezegun
pytz==2019.3              # via babel, django
regex==2020.2.20          # via black
requests==2.23.0          # via django-anymail, sphinx
s3transfer==0.3.3         # via boto3
six==1.14.0               # via astunparse, django-anymail, freezegun, packaging, python-dateutil, sphinxcontrib-httpdomain, tox, virtualenv
snowballstemmer==2.0.0    # via sphinx
sphinx-rtd-theme==0.4.3   # via -r requirements.in
sphinx==3.0.3             # via -r requirements.in, sphinx-rtd-theme, sphinxcontrib-httpdomain, sphinxcontrib-httpexample
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-httpdomain==1.7.0  # via -r requirements.in, sphinxcontrib-httpexample
sphinxcontrib-httpexample==0.10.3  # via -r requirements.in
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
toml==0.10.0              # via black, tox
tox==3.14.5               # via -r requirements.in
typed-ast==1.4.1          # via black
urllib3==1.25.8           # via botocore, requests
uwsgi==2.0.18             # via -r requirements.in
virtualenv==20.0.14       # via tox
wcwidth==0.1.9            # via pytest
wheel==0.34.2             # via astunparse
